+++ pkg/analysis_server/test/src/services/correction/fix/data_driven/flutter_use_case_test.dart
@@ -437,6 +437,51 @@ void f(IconButton button) {
 ''');
   }
 
+  Future<void>
+  test_flutter_to_material_package_IconButton_deprecated_export() async {
+    newFile('$workspaceRootPath/p/lib/flutter.dart', '''
+class Widget {
+}
+''');
+    newFile('$workspaceRootPath/p/lib/material.dart', '''
+@deprecated
+library;
+class IconButton {
+  const IconButton();
+}
+''');
+
+    newFile('$workspaceRootPath/p2/lib/material.dart', '''
+class IconButton {
+  const IconButton();
+}
+''');
+
+    writeTestPackageConfig(
+      config: PackageConfigFileBuilder()
+        ..add(name: 'p', rootFolder: getFolder('$workspaceRootPath/p')),
+    );
+
+    addPackageDataFile('''
+version: 1
+transforms:
+  - title: 'Replace by material from package p2'
+    date: 2023-11-09
+    library: 'package:p/material.dart'
+    changes:
+    - kind: 'replacedBy'
+      newLibrary: 'package:p2/material.dart'
+''');
+
+    await resolveTestCode('''
+export 'package:p/material.dart';
+''');
+
+    await assertHasFix('''
+export 'package:p2/material.dart';
+''');
+  }
+
   Future<void> test_flutter_to_material_package_IconButton_removed() async {
     newFile('$workspaceRootPath/p/lib/flutter.dart', '''
   class Widget {